Moved odds strategies to their own module, added light side and dark side odds strategies.
