+++ crapssim/strategy/simple_bet.py
@@ -3,7 +3,8 @@ is allowed."""
 import enum
 import typing
 
-from crapssim.strategy.core import OddsAmountStrategy, Strategy
+from crapssim.strategy.core import Strategy
+from crapssim.strategy.odds import OddsAmountStrategy
 
 from crapssim.bet import PassLine, DontPass, Bet, DontCome, Place, HardWay, Field, Three, Two, \
     Any7, Yo, Boxcars, Fire
